Fix compilation of equality checks for 'True' and 'False' patterns
Fix the compilation of the equality checks for patterns using 'Bool' values, which can be used, for example, in 'case ... of' expressions: These boolean values are implemented as choice type tags. The previous version compiled the pattern matching in a way that always expected the generic structure of choice type values.
This commit adapts the pattern matching to consider that we replace 'True' and 'False' with specific values in other places.