Reduce runtime expenses by reducing DecodeAndEval at compile-time
Adapt the compilation of Pine expressions to C# to emit representations with fewer runtime costs: Reduce instances of DecodeAndEvaluate expressions for which the decoded expression is trivial to determine at compile-time.

An outlook: There are additional cases where we want to reduce instances of DecodeAndEvaluate. However, these will require more expansive static analysis to prove that a particular expression will evaluate to a specific value within a larger context. With this commit, we pick some low-hanging fruits by handling the subset where local information is enough to prove that.