Speed up programs optimizing common subexpression elimination
Reduce runtime expenses for and speed up Elm and Pine programs with an optimized version of common subexpression elimination: Avoid unecessary evaluations of bindings we don't need in a specific application: Place them closer to the consuming site so they are only evaluated in the branches where necessary.
Also, search for subexpressions for CSE only in the necessary cases, at the root of an expression function, or when descending into the branches of a conditional expression.
Also, merge the model for compiled expression dependencies into the general record describing a compiled expression to simplify transport and transformations.
Expand the EnvironmentConfig type to model the available let bindings via a chain of scopes instead of placing all let bindings in a single list.