Simplify the core Elm compiler module by separating parsing
Prepare to move the most expensive portion of compiling Elm code to a more efficient implementation by moving the part for parsing Elm module syntax from a string into a separate module. The overhead for parsing is relatively low since it only requires a single module file as a context. Therefore, the parsing part is not the text target for optimization.