Add asynchronous logging to investigate sporadically hanging test runs
Recent runs on Azure Pipelines and GitHub Actions sporadically did not complete and were canceled after 6 hours and 1 hour, respectively.
The logs from GitHub showed that the step running tests via `dotnet test` did not complete, and the last started test was `TestElmInteractiveScenarios`
Therefore, add asynchronous logging to track the starts and completions of individual submissions in the test scenarios.