Fix bug in compiling recursive Elm functions in let blocks
Fix a bug that caused compilation to fail for some programs containing a recursive function in a let block.
Besides the fix on the productive side, this commit adds regression test scenarios that failed with the previous version of the compiler.

The wrapping of functions in a 'FunctionApplicationExpression' as done in 'closurizeFunctionExpressions' broke assumptions made in 'preprocessExpression': How we expected to be able to retrieve the inner expression (via 'parseFunctionParameters') did not work anymore after such a wrapping was applied.
To fix the compilation for these scenarios, this commit changes the transformations done for closures to move the additional application expression further to the consuming side for named functions.