Support focus on codegen challenges with separate module 🌲
Slice the elephant 🐘 ➡️ 🔪
Split the Elm compiler module - one more time 🎵 - Move the functionality for the later processing stages into the new module 'FirCompiler' 🎄
Recently, I attempted to simplify and clean up the code to compile from Elm to Pine. After having spent several hours on this project, I still had not succeeded in finding a better variant for this transformation. (I wanted to simplify or entirely remove the lifting, but understanding this and the related parts, like the implementations for closures and recursion, was too complex). One more time, the part mapping from the current intermediate language to Pine ('emit' functions) seemed more complicated than the earlier stage mapping from Elm to the intermediate language. After this experience, it seems more attractive to target this intermediate expression model when implementing mappings for other front-end languages (e.g., Roc, Gren, Cara) and reuse the implementation of the later stage.