Fix bug in Fir compiler and simplify compilation and generated code
Fix a bug in the Fir to Pine compiler that caused a mismatch in parameter/argument count in some scenarios with nested declaration blocks. This commit also adds a regression test that failed with the previous compiler version.
Since understanding how the lifting works turned out to be difficult, the winning approach does away with the lifting and flattening of declaration blocks entirely and instead introduces a new way of aggregating the environment for nested declaration blocks.