Evaluate some expressions in let-blocks earlier
Change the code emitted for declaration blocks to eagerly evaluate the subset of declarations that has zero parameters and does not depend on other declarations from the same block. This change fixes an issue that caused exponential growth in processing time in cases where such a declaration contained a recursive call and was used multiple times.