Work around problems with asynchronous SSL certificate arrival
Work around the problem with ASP.NET crashing because SSL certificates arrive ordered via LetsEncrypt arrive after starting the app: Disable HTTPS for a first start, if we find the certificate is not available. Later, if the HTTPS certificate has arrived, restart the ASP.NET host with the HTTPS URLS.
For discussion of this issue, see:

+ <https://github.com/ffMathy/FluffySpoon.AspNet.EncryptWeMust/issues/151>
+ <https://github.com/natemcmaster/LettuceEncrypt/issues/293>
+ <https://github.com/dotnet/aspnetcore/issues/26258>
+ <https://github.com/dotnet/aspnetcore/issues/45801>

Also, update the LetsEncrypt library and Certes to integrate various  recent upstream improvements.