Fix bug in special codegen for kernel 'skip' and 'list_head' sequence
Restore consistency with Pine kernel behavior, fixing a bug in the code generated as a more compact representation of a sequence of 'list_head' and 'skip' applications: In the previous version, the use of a path corresponding to a single 'list_head' application on an empty list would result in a .NET null reference. This commit applies the empty list default to restore symmetry with the behavior of the 'list_head' kernel function.