Improve yield in code analysis for recursive functions
Adapt code analysis to improve the environment types reported for expressions participating in recursion: For expressions collected on recursion paths, no longer report the environment type computed on entry at the root but the environment type resulting after returning from descendants. These are the environment types we actually need, so this set of analysis results leads to more specialized representations in the emit stage. So, the outcome is a reduction of detours via generic representations at runtime.