Fix potential ambiguities in integer encoding
Avoid mapping from Pine value to integer in cases where the encoding in the other direction would no longer result in the same value. Sometimes, the relaxed version from the previous version would be OK. One example of such acceptable use is when a kernel function consumes the integer. However, such a non-roundtripping conversion could lead to corruption on interfaces used for compression or optimization.