Refactor to reduce the compiled size of the Elm compiler
Find a refactoring that massively reduces the size of the compiled Elm compiler - by more than 80%. After running experiments this week, I found a problem with compiling the '++' operator, causing exponential growth of the emitted code. The refactoring in this commit replaces usages of this operator with more specific functions, 'String.join' and 'List.concat', depending on the argument type.