Use type inference in Elm compiler to optimize operator application
Use simple type inference to search for optimized compilation of operator application.
For example, in a sequence of ++ operators: If any element is a string literal, within the rules
of Elm, all other elements also must be of type String. (This is in contrast to other programming languages that would dynamically convert some other types to String)
This constraint of types allows for the emission of a specialized representation, reducing compiled code size and runtime overhead.