Fix display for other file types besides Elm modules
The previous implementation only considered Elm modules in the setup of the Monaco editor and surroundings, resulting in suboptimal visuals.
The changes in this commit expand support to other file types, specifically JSON, XML, Markdown, HTML, and CSS.
Also, display the 'Format' and 'Compile' buttons only when these are usable on the file currently open in the editor.