Fix a bug in the compilation of record updates for multiple fields
Fix a defect in the compilation of Elm record updates that caused some updates to crash at runtime: The code emitted by the previous version only worked if the updated fields appeared sorted by name in the original Elm syntax. This commit changes the compiler so that the updates are sorted accordingly before handing the list to the generic record update function.