Improve readability with specialized emission on predictable parse error
Improve the readability of the intermediate code emitted by the Pine compiler by emitting specialized code in cases where we can predict a parsing error instead of defaulting to emit the general representation to call the interpreter to evaluate any expression.
The Elm compiler uses the pattern of emitting predictable parsing errors to encode runtime checks with custom error messages. The most common usage currently might be the message "Error in case-of block: No matching branch."