Fix stack overflows in Pine interpreter with custom stack implementation
The previous version of the Pine interpreter consistently crashed with a stack overflow when running the Elm compiler to compile all the Elm compiler modules. In that workload, it compiled the first 13 Elm modules successfully but crashed when compiling the `Json.Decode` module.
The new interpreter introduced with this commit uses an explicit stack model instead of reusing the stack of the underlying dotnet runtime.