Improve performance with instruction fusion of `list_head (skip(a, b))`
Reduce runtime expenses with an optimization specialized for the form combining the kernel functions `skip` and `list_head` into a pattern like `list_head (skip(a, b))`: Expand the preprocessing in the interpreter to fuse this pattern into a single dedicated expression node. One benefit of this fusion is skipping the creation of the intermediate shorter list object.