Expand automated test to compile Elm compiler on Pine
Expand the test program that uses the Pine engine to execute the Elm compiler compiling the Elm compiler and checks if the resulting module values equal the output from the Elm compiler running on JavaScript/V8.
Tune the training scenario to speed up the profiling and code analysis.
For this test, disable code analysis and compilation of Pine expressions to .NET and switch to memoization as the only optimization for now.
Add detailed reports on each compiled Elm module, listing each declaration separately to support inspection and comparison of Elm compiler versions.