Improve consistency in the splitting of conditional expressions
Adapt the compilation of conditional expressions to instructions to apply the lifting of contained parse&eval expressions also for the condition part, as we were already doing for the branches.
Also, simplify the conditional jump instruction model so that it always implicitly uses the result from the last instruction as the condition value.