Tune implementation and data structures in Elm compiler for efficiency
Refactor the implementation of the Elm compiler and some functions in the core libraries to improve efficiency, mainly during compilation. Some of the optimizations in this commit result from the reduction of partial applications. These changes follow observations last month revealing that the compiler spent a lot of instruction cycles in the generic partial application wrapper we emit in the Elm compiler. The Fir compiler already contains some smartness to emit more direct calls when the shape of the function is known early, avoiding the expensive partial application. However, it does not apply this statically informed specialization when invoking a function from a parameter. Popular examples of such cases are `List.map` and `List.concatMap`. The refactorings in this commit achieve a reduction of the number of total partial application iterations in two different ways. One way is to use specialized functions so that we no longer need to apply a function from an argument. Another way is combining multiple parameters into a single parameter (as done, for example in `resultListIndexedMapCombine`)
Another change in this commit that resulted in significantly improved efficiency is the reduction of operations on Elm records. One reason records are currently expensive is that we don't do any type inference on them and, therefore, emit generic code for getting or updating field values. The generic record update works with different record types and scans the whole list of fields to find the right place(s) to insert new field values. Switching to choice type tags (as done, for example, in `EnvironmentFunctionEntry`) means we now avoid these scans and use positional addressing.