Improve runtime performance with better-specializing compilation
Expand code analysis and reduction for environment classes. The automated test added with this commit shows how we now achieve more than twice the efficiency for many common Elm record access operations. This efficiency gain results from code analysis and compilation improvements introduced with this commit.

The previous approach to code analysis aimed at identifying the parts of the environment used to parse expressions. Looking closer at optimizing execution in PineVM this month, it became clear we want to add more parts of the environment constraints collected for specialized compilation. Expanding the environment constraint to cover parts that conditions of conditional expressions depend on allows us to resolve the condition when compiling the specialized representation. This compile-time reduction saves us the check at runtime and shrinks the code size because we eliminate some branches.

But how do we expand the code analysis to cover the condition expressions? Initially, I expected to expand the existing code, adding something similar to what we did so far to identify the subexpressions used in parsing. However, this commit adds an approach to collecting the environment constraints that is entirely different from the one used so far. Instead of looking into the contents of the expression and recursively descending to collect all relevant paths, this new approach does not look into the expression. Instead, it generates classes only by computing intersections of the different concrete environments observed for a given expression. It then computes a score for each class by counting how many of the observed concrete environments match the class. This approach seems preferable for now because it is much simpler. However, the concrete implementation still needs to be refined.