Refine environment classes to filter out superfluous items
This commit also adds a new automated test with scenarios of Elm record updates to illustrate the new capability for refining environment classes.
The simplification of environment classes added with this commit is the first step to refining the set of environment classes we hand over to the execution engine for specialized compilation.