Optimize further - more inlining when compiling for environment classes
This commit also adds an automated test to illustrate a scenario that benefits from the change in the policy for inlining parse&eval expressions. That new test case shows the complete inlining of the function parameter to `List.map`
Since the item mapping function is so simple in this case (`Tuple.first`), we expect the engine (with the given configuration) to create a specialized variant of the `List.map` function with this item mapping function inlined.
Therefore, the test expects only one invocation per item of the input list.

Another notable change in this commit is how we aggregate the set of environment classes: We no longer remove more general classes when adding a new class. That rule in the previous version broke the optimization in the new test scenario using 'List.map'. Removing this rule enabled optimal performance for that scenario and improved overall performance in scenarios with larger applications and workloads.

When optimizing `List.map` (or its recursive helper function) (or `List.foldx` for example), better also inline the application of the generic partial application used with the function parameter.
That application is conditional (we only end up in this branch if the list is not empty), but we want to inline that to eliminate the generic wrapper for the function application and inline the parameter function directly.
Thus, the new rule also enables inlining under conditional expressions unless it is recursive.