Rearrange references forwarded when emitting function application
For the emission of function applications, move the closure captures from the first item in the environment list into the second element together with the function arguments. This change brought significant savings in emitted module size and instructions executed at runtime. It would be nice to repeat this trick for the other case where we compose new environments, the emission of declaration blocks (originating, for example, from let-blocks)