Rewrite symbol resolution for readability and efficiency
Change the internal representation of references to not fuse module name and declaration name into a flat string anymore. Instead, we expand the reference model in the Fir compiler to hold the module name separately. One benefit of this change is saving operations when building the collection(s) representing the imports.
Expand the compilation to map local references to canonical names earlier (according to the import statements found in the current module)