Optimize Pine expression encoding for runtime
So far, I expected to develop the standard expression encoding into a form completely representable in a file system tree. With such an encoding, we could take a tree from the file system and directly feed it into Parse&Eval. However, I changed my mind about the tradeoffs, and this commit changes the encoding in the opposite direction, optimizing for compactness and symmetry with the Elm value encoding.
If we still want to (re)use file systems to model syntax trees, we will add a mapping for such a form later, meaning there would be another hash code for that encoding.