Explore precompiled builtin for List.foldl and List.concatMap
In contrast to others covered before as precompiled, these take functions as arguments.
One advantage of this change is we save the lookup/resolution of the adaptive partial application for each item in the input list. However, measurements of time spent indicate that the savings from optimizing `List.concatMap` and `List.foldl` this way is small despite reducing the count of overall function applications by more than 15 percent.