Clean up expression model used in VM and parse cache
Clean the expression type used in the VM, removing the delegating expression variant. We used the delegate variant to link compiled representations. But for linking (pre)compiled representations as done now, we only need a redirection branch on invocation (Parse&Eval), not as a possibility on every subexpression. This commit changes the interface to inject redirections accordingly, constraining it to cases of Parse&Eval or similar invocations.
After the change in overriding invocations, we no longer have any use for configuring/overriding the parsing of expressions in the VM, so this commit reduces the customizability of parsing in the VM.