Improve efficiency by reusing more instances of common values
Expand the reuse of instances to reduce heap allocations and speed up equality checks.
Besides expanding the set of values for which we reuse instances, this commit also rewrites the implementation to initialize these collections to improve the readability of the program code.
The continuous incremental growth of collections of instances to reuse and the corresponding implementations for their initialization was a lesson in how to (not) deal with type initializers. The earlier implementation used separate initialization methods for individual collections or pairs of maps, and the program invoked these methods via separate assignments of static declarations/type initializers. The implementations to construct these collections introduce cyclic dependencies, and these cycles are necessary in some sense. However, the distribution over multiple methods and resolution of the cycles via the reentry rules for type initializers meant it was difficult to understand the resulting sequence at runtime. This commit consolidates the initialization of these collections and maps into a single method to make the execution order obvious.