Fix record field order when constructing via alias declaration
In Elm, a type alias declaration for a record type also implicitly creates a constructor function to create such a record. This implicitly declared function has the same name as the alias and takes the field values in the same order as they appear in the alias declaration.
However, the compiled representations of equality checks and record updates expect the record fields to be ordered alphabetically. Therefore, this commit implements a mapping from the constructor argument order to the canonical order of the fields.