Speed up automated tests by moving compilation from JavaScript to Pine
Move the compilation of Elm programs in the automated tests from the JavaScript-based implementation to the Pine-based implementation. This change reduced the execution time for the combination of core scenarios and kernel scenarios to less than half.
The part parsing the interactive submission from a string into a syntax tree still runs on JavaScript and will be migrated in a future commit.