Begin bundling precompiled Elm compiler for faster startup
Expand the prebuild task to bundle the compiled representation of the default Elm compiler. This bundling enables faster startup of the application, benefiting scenarios of a first start on a machine where no cache of a previous compilation is found. In GitHub Actions, this change reduced runtimes by ten minutes for the main Action running all tests on Linux, Windows, and macOS environments.
Besides expanding the prebuild task, this commit also expands the CLI, adding a command to compile a specific set of Elm Interactive environments into a file for bundling. Exposing this functionality via the CLI means that the prebuild task can (in the future) use the binary from a previous release instead of repeating the bootstrapping via JavaScript as done so far. This should help speed up the test runs in GitHub Actions since this route for compilation is already faster than using V8.