Update encoding from base64 to `Bytes` in `Platform.WebService`
Switch the encoding for HTTP request and response bodies from base64 to type `Bytes` from the Elm core libraries. This interface makes for a better API and is more efficient at runtime. The earlier use of the base64 encoding was an adaption to the use of JavaScript engines, which, in turn, were replaced this year.