Switch standard string encoding from a list of chars to a blob
Switch the encoding of strings used to model (names in) file system trees and expression tags to a sequence of bytes instead of a list of characters. This commit also switches tag labels for choice types emitted by the Elm compiler to the new string format. Using the same string model for Elm strings seems preferable, so that will probably follow in a future change.