Add general tail call optimization in IR VM bytecode interpreter
Implement a general form of tail call optimization, replacing the current stack frame when we encounter a call in the tail position. (For the special case of a function calling itself, the plan is to add another optimization in the future, translating these into local looping)
The primary motivation for adding this optimization now is to fix excessive memory usage: With the virtual stack model, we don't have a problem with stack overflows. However, with the current heavyweight stack frame representation, these reducible frames sometimes use many gigabytes of memory.