Fix bug in preparing root from function record in code analysis
Hunting this bug cost me more than five hours; it was a confusing journey. While inspecting intermediate results, I noticed that some of the parse&eval instances in the analyzed expressions appear to be mixed up, having the environment from one instance but pointing to the encoded expression that should appear in another instance. However, I first suspected that the mix-up originated from a bug in the inlining and reduction process happening in `ParseAsStaticMonomorphicProgram`, before finally discovering that the problem lies right at the root. In the case of the `hexStringToInt` scenario added with this commit, there are two environment function entries for `hexStringBytesToInt` and `charToHex`. However, the function record in the declaration `hexStringToInt` has them ordered differently than `hexStringBytesToInt`.