dispatch from .NET into the compiled representation of Pine-programs
This commit adds the implementation to dispatch invocations from .NET into the (JIT)-compiled representation of Pine-programs compiled via the new static-monomorphized-program route. It expands the generation of C# code to include a dispatcher module. This dispatcher module contains a dictionary that maps functions to .NET methods that implement these functions.
Also in this commit: Fixes for the emission to C#, for example, adding correct qualifiers on type references according to the configured available namespaces/usings.