Improve runtime efficiency for sequence construction by lazy evaluation
This commit introduces local lazy evaluation to improve runtime efficiency in scenarios with many nested applications of the 'concat' kernel function. In cases where we build lists by appending or prepending one item at a time, the naive representation that immediately evaluated the 'concat' applications resulted in a runtime quadratic in the number of items, since evaluation of each 'concat' allocates a new blob or list. To avoid these exponential costs, we introduce a representation that records 'concat' operations for lazy evaluation.
This optimization is similar to 'ConsString' in V8. One difference is that we don't limit this to strings or blobs; we apply it to lists, too.
Another difference of the laziness as introduced in this commit is that we do some of the work upfront: we analyze the shape of functions to decide whether to use such a representation for a local, and force evaluation when it escapes the current function. To achieve maximum performance long-term, the goal is to eventually allow cross-function construction and use both statically and dynamically decided laziness.
Long-term, adding proofs to distinguish between blobs and lists might further improve runtime efficiency (see roadmap items).