Better runtime efficiency by expanding use of optimized representation
Expand the use of the internal value representation to go cross-stack frame, so that we force expensive evaluation/materialization less often.

Note: Our engine has a special handling of tail calls: It replaces the current stack frame, which means the method 'returning' a value from a stack frame returns right to the last non-tail call. As a consequence, the part responsible for memoizing the result is only invoked once for a whole chain of tail-call recursions. So the forcing of an evaluation of the returned value (for the purpose of building a cache-key) also only happens in that case.
Nonetheless, in the future, this can become even more efficient by avoiding more of these building of list instances for cache keys.