Switch IR to the optimized function parameter model
Switch the intermediate representation in the interpreter to use the specialized parameter model instead of the generic environment value as a single parameter.
The specialized model of inputs to each stack frame enables us to hash these inputs with reduced costs: We can now reuse hashes for frequently appearing arguments to build the overall hash for the combination of inputs.