point out primary purpose
A bit of history: At an earlier stage, the idea was to design the expression encoding so that we could just use a file system tree with directory names and filenames to build a valid Pine expression. This way, you could reuse the already tree-shaped filesystem API to model a syntax tree, to bootstrap a language before you have implemented a parser. I gave up that feature to optimize the expression encoding for density.